Get nodepool doing something
This gets nodepool to the point where it's trying to build images. Build
fails for me around the tail end of it, trying to mess with loopback
stuff. But I want to push this up so others can play with it.

Another new secret, clouds.yaml to interact with an OpenStack cloud.